// frontend/src/api/generated/models/operatorAdminResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Manufacturing Information System API
 * API untuk aplikasi core MIS.
 * OpenAPI spec version: 1.0.0
 */
import type { DeletionEligibility } from './deletionEligibility.ts';

export interface OperatorAdminResponse {
  id: string;
  user_id: number;
  username: string;
  first_name: string;
  last_name: string;
  full_name: string;
  email: string;
  operator_type: string;
  status: string;
  supervisor_id: string | null;
  supervisor_name: string;
  location: string;
  phone: string;
  account_is_active: boolean;
  work_is_active: boolean;
  deletion_eligibility: DeletionEligibility;
}